requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
class-registry==2.1.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.1
django-environ==0.12.0
django_celery_results==2.6.0
djangorestframework==3.16.1
drf-yasg==1.21.14
filters==1.3.2
gunicorn==25.0.1
inflection==0.5.1
kombu==5.6.2
packaging==26.0
pika==1.3.2
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.5.3
